backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0

# ML/AI dependencies - Updated for Python 3.13 compatibility
scikit-learn>=1.4.0
pandas>=2.1.3
numpy>=1.25.2
matplotlib>=3.8.2
seaborn>=0.13.0
joblib>=1.3.2

# Optional: For advanced ML features
# tensorflow==2.15.0
# torch==2.1.1
